requirements.txt
openai
whoosh
python-dotenv   
streamlit